refactor: split file for exportConfig
ONE-vscode-DCO-1.0-Signed-off-by: Yeseo Kang <yeeybook@naver.com>
