+++ src/configuration-settings/ConfigurationSettingsPanel.ts
@@ -2,6 +2,7 @@ import { request } from 'http';
 import * as vscode from 'vscode';
 import {getNonce} from '../getNonce';
 import {importConfig} from './Dialog/ImportConfigDialog'
+import { exportConfig } from './Dialog/ExportConfigDialog';
 // import { PathDialog } from './PathDialog';
 
 export class ConfigurationSettingsPanel {
@@ -53,24 +54,24 @@ export class ConfigurationSettingsPanel {
     ConfigurationSettingsPanel.currentPanel = new ConfigurationSettingsPanel(panel, extensionUri);
   }
 
-  private exportConfig(){
-    const ConfigPareser = require('configparser');
-    const config = new ConfigPareser();
+  // private exportConfig(){
+  //   const ConfigPareser = require('configparser');
+  //   const config = new ConfigPareser();
     
-    // Adding sections and adding keys
-    config.addSection('User');
-    config.set('User', 'token', 'some value');
-    config.set('User', 'exp', 'some value');
+  //   // Adding sections and adding keys
+  //   config.addSection('User');
+  //   config.set('User', 'token', 'some value');
+  //   config.set('User', 'exp', 'some value');
 
-    // With String Interpolation, %(key_name)s
-    config.addSection('MetaData');
-    config.set('MetaData', 'path', '/home/%(dir_name)s/');
-    config.set('MetaData', 'dir_name', 'me');
+  //   // With String Interpolation, %(key_name)s
+  //   config.addSection('MetaData');
+  //   config.set('MetaData', 'path', '/home/%(dir_name)s/');
+  //   config.set('MetaData', 'dir_name', 'me');
 
-    // config.write('my-cfg-file.cfg');
-    console.log('successfully wrote file!!!')
-    return config;
-  }
+  //   // config.write('my-cfg-file.cfg');
+  //   console.log('successfully wrote file!!!')
+  //   return config;
+  // }
 
   private constructor(panel: vscode.WebviewPanel, extensionUri: vscode.Uri) {
     this._panel = panel;
@@ -140,29 +141,11 @@ export class ConfigurationSettingsPanel {
           );
         break;
         case 'exportConfig':
-          const optionsForExportDialog: vscode.SaveDialogOptions = {
-            filters: {
-              'allFiles': ['*']
-            }
-          };
-          vscode.window.showSaveDialog(optionsForExportDialog).then(fileUri => {
-            if (fileUri) {
-              this.exportConfig().write(fileUri.path + '.cfg');
-              console.log('Selected file!!!: ' + fileUri.path + '.cfg');
-              }
-            );
-          break;
-        case 'alert': 
-          vscode.window.showErrorMessage(data.text);
-          break;
-        case "importConfig":
-            const newWebview = this._panel.webview;
-            newWebview.html = this._getHtmlForWebview(newWebview);
-            importConfig(newWebview);
-            break;
-        }
-      });
-    }
+          exportConfig();
+        break;
+      }
+    });
+  }
 
   private _getHtmlForWebview(webview: vscode.Webview) {
     // And the uri we use to load this script in the webview
